+++ website/siteConfig.js
@@ -53,6 +53,7 @@ const siteConfig = {
     // For no header links in the top nav bar -> headerLinks: [],
     headerLinks: [
         {href: '/airframe/docs/', label: 'Docs'},
+        {href: '/airframe/docs/release-notes', label: 'Release Notes'},
         {href: 'https://github.com/wvlet/airframe/', label: 'GitHub'},
     ],